experiments/cloud_functions_copy/requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-logging
